+++ compiler/rustc_metadata/src/rmeta/parameterized.rs
@@ -104,18 +104,18 @@ trivially_parameterized_over_tcx! {
     rustc_hir::attrs::StrippedCfgItem<rustc_hir::def_id::DefIndex>,
     rustc_hir::attrs::lang_items::LangItem,
     rustc_hir::def::DefKind,
-    rustc_hir::def::DocLinkResMap,
     rustc_hir::def_id::DefId,
     rustc_hir::def_id::DefIndex,
     rustc_hir::definitions::DefKey,
     rustc_index::bit_set::DenseBitSet<u32>,
-    rustc_middle::metadata::AmbigModChild,
-    rustc_middle::metadata::ModChild,
     rustc_middle::middle::codegen_fn_attrs::CodegenFnAttrs,
     rustc_middle::middle::debugger_visualizer::DebuggerVisualizerFile,
     rustc_middle::middle::deduced_param_attrs::DeducedParamAttrs,
     rustc_middle::middle::exported_symbols::SymbolExportInfo,
     rustc_middle::middle::lib_features::FeatureStability,
+    rustc_middle::middle::resolve::AmbigModChild,
+    rustc_middle::middle::resolve::DocLinkResMap,
+    rustc_middle::middle::resolve::ModChild,
     rustc_middle::middle::resolve_bound_vars::ObjectLifetimeDefault,
     rustc_middle::mir::ConstQualifs,
     rustc_middle::mir::ConstValue,
