+++ compiler/rustc_passes/src/lang_items.rs
@@ -13,8 +13,9 @@ use rustc_crate_store::ExternCrate;
 use rustc_hir::Target;
 use rustc_hir::attrs::lang_items::{GenericRequirement, LangItem, LanguageItems};
 use rustc_hir::def_id::{DefId, LocalDefId};
+use rustc_middle::middle::resolve::ResolverAstLowering;
 use rustc_middle::query::Providers;
-use rustc_middle::ty::{ResolverAstLowering, TyCtxt};
+use rustc_middle::ty::TyCtxt;
 use rustc_span::{Span, Symbol, sym};
 
 use crate::diagnostics::{DuplicateLangItem, IncorrectCrateType, IncorrectTarget};
