+++ src/librustdoc/passes/lint/redundant_explicit_links.rs
@@ -3,8 +3,9 @@ use std::ops::Range;
 use rustc_ast::NodeId;
 use rustc_errors::{Diag, DiagCtxtHandle, Diagnostic, Level, SuggestionStyle};
 use rustc_hir::HirId;
-use rustc_hir::def::{DefKind, DocLinkResMap, Namespace, Res};
+use rustc_hir::def::{DefKind, Namespace, Res};
 use rustc_lint::Applicability;
+use rustc_middle::middle::resolve::DocLinkResMap;
 use rustc_resolve::rustdoc::pulldown_cmark::{
     BrokenLink, BrokenLinkCallback, CowStr, Event, LinkType, OffsetIter, Parser, Tag,
 };
